Changes Enrollment form
