+++ frontend/src/Components/Enrollment/CampaignEnrollment.css
@@ -1,4 +1,4 @@
-
+/* 
 *{
     padding: 0;
     margin:0;
@@ -27,4 +27,4 @@ button
 body
 {
     background-image: linear-gradient(to right,	white,	pink);
-}
+} */
